demo điều hướng

+++ app/src/main/java/com/example/jcj/learningskid/SQLiteHelper.java
@@ -20,18 +20,33 @@ public class SQLiteHelper extends SQLiteOpenHelper {
 
     }
 
+    public SQLiteHelper(Context context, String dbName, int version) {
+
+        super(context, dbName, null, 1);
+        this.dbName = dbName;
+    }
     @Override
     public void onCreate(SQLiteDatabase database) {
 
         String CREATE_TABLE = "CREATE TABLE IF NOT EXISTS " + tableName + " (" + Table_Column_ID + " INTEGER PRIMARY KEY, " + Table_Column_1_Name + " VARCHAR, " + Table_Column_2_Email + " VARCHAR, " + Table_Column_3_Password + " VARCHAR)";
         database.execSQL(CREATE_TABLE);
 
+        String createUserEnglishContentGrammar = "CREATE TABLE IF NOT EXISTS UserEnglishContentGrammar(" +
+                "id integer primary key autoincrement," +
+                "name text not null," +
+                "content text)";
+        database.execSQL(createUserEnglishContentGrammar);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + tableName);
-        onCreate(db);
+        if(newVersion != oldVersion){
+            db.execSQL("DROP TABLE IF EXISTS " + tableName);
+            String sql = "DROP TABLE IF EXISTS UserEnglishContentGrammar";
+            db.execSQL(sql);
+            onCreate(db);
+
+        }
 
     }
